# docs/swagger.yaml
definitions:
  api.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.AuthenticateUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  dto.RegisterUserRequest:
    properties:
      address:
        type: string
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.AuthenticateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Authenticate user
      tags:
      - auth
  /api/register:
    post:
      consumes:
      - application/json
      description: Register a new user with default role
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register User
      tags:
      - auth
  /api/users:
    get:
      consumes:
      - application/json
      description: Get all users with their roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Get all users
      tags:
      - users
swagger: "2.0"